Simplify tempcity
Instance attributes are not really useful unless you plan to use
them in other methods. Local variables are both faster and easier to read.

Naming variables with single letters (e.g. "e") is a bit ugly and in
certain cases (like this one) can be even dangerous (when you reuse
the same name in the same scope)...